tests: add coverage tests; refactor a bit
