+++ .gitignore
@@ -99,4 +99,8 @@ listfile.csv
 output/
 outputs/
 
+tests/.coverage
+tests/htmlcov
+tests/report.html
+
 !.notempty
