+++ wowdump/dumputil.py
@@ -9,6 +9,40 @@ import logging
 from kaitaistruct import BytesIO, KaitaiStream, KaitaiStruct
 from ppretty import ppretty
 
+
+# "kaitai type"
+def ktype(v):
+    logger = logging.getLogger()
+    if inspect.isclass(v):
+        logger.debug("var is type class, skipping")
+        return "skip"
+
+    if inspect.ismethod(v) or inspect.isbuiltin(v):
+        logger.debug("var is is method or builtin, skipping")
+        return "skip"
+
+    if isinstance(v, type):
+        logger.debug("var is defines a datatype, skipping")
+        return "skip"
+
+    if isinstance(v, list):
+        logger.debug("var is type list")
+        return "list"
+
+    if isinstance(v, KaitaiStruct):
+        logger.debug(f"var is a kaitai type {type(v)}")
+        return "kaitai"
+
+    if type(v) in [int, float, str, bool]:
+        logger.debug("var is a base type")
+        return "base"
+
+    # otherwise is a dict type
+    # FIXME: I think dict type, anyhow
+    logger.debug("var is (probably) a dict type")
+    return "dict"
+
+
 def kttree(obj, path: str=""):
     r = {}
     lgkttree = logging.getLogger("kttree")
@@ -76,3 +110,55 @@ def kttree(obj, path: str=""):
 
     # debug(f"out:  returning {r}")
     return r
+
+
+def whatis(obj):
+    objis = []
+
+    if inspect.ismodule(obj):
+        objis.append("module")
+    if inspect.isclass(obj):
+        objis.append("class")
+    if inspect.ismethod(obj):
+        objis.append("method")
+    if inspect.isfunction(obj):
+        objis.append("function")
+    if inspect.isgeneratorfunction(obj):
+        objis.append("generatorfunction")
+    if inspect.isgenerator(obj):
+        objis.append("generator")
+    if inspect.iscoroutinefunction(obj):
+        objis.append("coroutinefunction")
+    if inspect.iscoroutine(obj):
+        objis.append("coroutine")
+    if inspect.isawaitable(obj):
+        objis.append("awaitable")
+    if inspect.isasyncgenfunction(obj):
+        objis.append("asyncgenfunction")
+    if inspect.isasyncgen(obj):
+        objis.append("asyncgen")
+    if inspect.istraceback(obj):
+        objis.append("traceback")
+    if inspect.isframe(obj):
+        objis.append("frame")
+    if inspect.iscode(obj):
+        objis.append("code")
+    if inspect.isbuiltin(obj):
+        objis.append("builtin")
+    if inspect.isroutine(obj):
+        objis.append("routine")
+    if inspect.isabstract(obj):
+        objis.append("abstract")
+    if inspect.ismethoddescriptor(obj):
+        objis.append("methoddescriptor")
+    if inspect.isdatadescriptor(obj):
+        objis.append("datadescriptor")
+    if inspect.isgetsetdescriptor(obj):
+        objis.append("getsetdescriptor")
+    if inspect.ismemberdescriptor(obj):
+        objis.append("memberdescriptor")
+
+    if isinstance(obj, KaitaiStruct):
+        objis.append("kaitaistruct")
+
+    return objis
